refactor: AttendanceStatus를 반환하는 메서드 AttendanceStatus로 위임
