+++ src/main/java/attendance/domain/HourMinute.java
@@ -6,6 +6,6 @@ public record HourMinute(int hour, int minute, AttendanceStatus attendanceStatus
     public static final int NULL_TIME = -1;
 
     public HourMinute(LocalDateTime localDateTime) {
-        this(localDateTime.getHour(), localDateTime.getMinute(), AttendanceChecker.checkAttendance(localDateTime));
+        this(localDateTime.getHour(), localDateTime.getMinute(), AttendanceStatus.checkAttendance(localDateTime));
     }
 }
