+++ src/test/java/attendance/domain/AttendanceCheckerTest.java
@@ -18,37 +18,6 @@ import org.junit.jupiter.params.provider.ValueSource;
 @DisplayName("출석 상태를 판별")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class AttendanceCheckerTest {
-    @ParameterizedTest
-    @CsvSource({
-            "10, 5, PRESENT",
-            "9, 5, PRESENT",
-            "10, 6, LATENESS",
-            "10, 30, LATENESS",
-            "10, 31, ABSENCE",
-            "12, 50, ABSENCE"
-    })
-    void 평일_출석_시간으로_출석_정보를_반환한다(int hour, int minute, AttendanceStatus expected) {
-        LocalDateTime attendanceTime = LocalDateTime.of(2024, 12, 24, hour, minute);
-        AttendanceStatus actual = AttendanceChecker.checkAttendance(attendanceTime);
-
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @ParameterizedTest
-    @CsvSource({
-            "13, 5, PRESENT",
-            "10, 5, PRESENT",
-            "13, 6, LATENESS",
-            "13, 30, LATENESS",
-            "13, 31, ABSENCE",
-            "15, 50, ABSENCE"
-    })
-    void 월요일_출석_시간으로_출석_정보를_반환한다(int hour, int minute, AttendanceStatus expected) {
-        LocalDateTime attendanceTime = LocalDateTime.of(2024, 12, 23, hour, minute);
-        AttendanceStatus actual = AttendanceChecker.checkAttendance(attendanceTime);
-
-        assertThat(actual).isEqualTo(expected);
-    }
 
     @ParameterizedTest
     @CsvSource({
